requirements.txt
# Core dependencies
llama-index>=0.9.0
python-dotenv>=1.0.0
pinecone-client>=2.2.4
openai>=1.3.0

# PDF processing
pypdf2>=3.0.0
pymupdf>=1.23.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Web interface
flask>=2.3.0
streamlit>=1.28.0

# Machine learning
langchain>=0.0.300
torch>=2.0.0
transformers>=4.30.0
torchvision>=0.15.0

# Visualization
matplotlib>=3.7.0
pillow>=10.0.0
